packages/api/requirements.txt
aiohttp==3.7.3
aredis==1.1.8
asyncpg==0.21.0
cryptography==3.3.1
hiredis==1.1.0
pyjwt==2.0.1
starlette==0.14.1
uvicorn==0.13.3
uvloop==0.14.0
